Final/back/requirements.txt
flask
celery
redis
Tkinter
#Python 3.11 
#Angular
TensorFlow
matplotlib
numpy
requests
scipy
ramdom 
#zipfile
pillow